+++ pkg/devfile/adapters/kubernetes/component/adapter.go
@@ -9,11 +9,10 @@ import (
 	"strings"
 	"time"
 
-	"github.com/openshift/odo/pkg/service"
-
 	"github.com/devfile/library/pkg/devfile/generator"
 	componentlabels "github.com/openshift/odo/pkg/component/labels"
 	"github.com/openshift/odo/pkg/envinfo"
+	"github.com/openshift/odo/pkg/service"
 	"github.com/openshift/odo/pkg/util"
 
 	appsv1 "k8s.io/api/apps/v1"
@@ -111,6 +110,7 @@ type Adapter struct {
 // Push updates the component if a matching component exists or creates one if it doesn't exist
 // Once the component has started, it will sync the source code to it.
 func (a Adapter) Push(parameters common.PushParameters) (err error) {
+
 	a.deployment, err = a.Client.GetKubeClient().GetOneDeployment(a.ComponentName, a.AppName)
 	if err != nil {
 		if _, ok := err.(*kclient.DeploymentNotFoundError); !ok {
@@ -160,6 +160,29 @@ func (a Adapter) Push(parameters common.PushParameters) (err error) {
 	}
 	s.End(true)
 
+	log.Info("\nUpdating services")
+	// fetch the "kubernetes inlined components" to create them on cluster
+	// from odo standpoint, these components contain yaml manifest of an odo service or an odo link
+	k8sComponents, err := a.Devfile.Data.GetComponents(parsercommon.DevfileOptions{
+		ComponentOptions: parsercommon.ComponentOptions{ComponentType: devfilev1.KubernetesComponentType},
+	})
+	if err != nil {
+		return errors.Wrap(err, "error while trying to fetch service(s) from devfile")
+	}
+	labels := componentlabels.GetLabels(a.ComponentName, a.AppName, true)
+	// create the Kubernetes objects from the manifest and delete the ones not in the devfile
+	needRestart, err := service.PushServiceFromKubernetesInlineComponents(a.Client.GetKubeClient(), k8sComponents, labels)
+	if err != nil {
+		return errors.Wrap(err, "failed to create service(s) associated with the component")
+	}
+
+	if componentExists && needRestart {
+		err = a.Client.GetKubeClient().WaitForPodNotReady(podName)
+		if err != nil {
+			return err
+		}
+	}
+
 	log.Infof("\nCreating Kubernetes resources for component %s", a.ComponentName)
 
 	previousMode := parameters.EnvSpecificInfo.GetRunMode()
@@ -183,21 +206,6 @@ func (a Adapter) Push(parameters common.PushParameters) (err error) {
 		return errors.Wrap(err, "unable to create or update component")
 	}
 
-	// fetch the "kubernetes inlined components" to create them on cluster
-	// from odo standpoint, these components contain yaml manifest of an odo service or an odo link
-	k8sComponents, err := a.Devfile.Data.GetComponents(parsercommon.DevfileOptions{
-		ComponentOptions: parsercommon.ComponentOptions{ComponentType: devfilev1.KubernetesComponentType},
-	})
-	if err != nil {
-		return errors.Wrap(err, "error while trying to fetch service(s) from devfile")
-	}
-	labels := componentlabels.GetLabels(a.ComponentName, a.AppName, true)
-	// create the Kubernetes objects from the manifest and delete the ones not in the devfile
-	err = service.PushServiceFromKubernetesInlineComponents(a.Client.GetKubeClient(), k8sComponents, labels)
-	if err != nil {
-		return errors.Wrap(err, "failed to create service(s) associated with the component")
-	}
-
 	a.deployment, err = a.Client.GetKubeClient().WaitForDeploymentRollout(a.deployment.Name)
 	if err != nil {
 		return errors.Wrap(err, "error while waiting for deployment rollout")
@@ -215,15 +223,21 @@ func (a Adapter) Push(parameters common.PushParameters) (err error) {
 		return err
 	}
 
+	ownerReference := generator.GetOwnerReference(a.deployment)
 	// update the owner reference of the PVCs with the deployment
 	for i := range pvcs {
 		if pvcs[i].OwnerReferences != nil || pvcs[i].DeletionTimestamp != nil {
 			continue
 		}
-		err = a.Client.GetKubeClient().UpdateStorageOwnerReference(&pvcs[i], generator.GetOwnerReference(a.deployment))
+		err = a.Client.GetKubeClient().UpdateStorageOwnerReference(&pvcs[i], ownerReference)
 		if err != nil {
 			return err
 		}
+	}
+
+	err = service.UpdateKubernetesInlineComponentsOwnerReferences(a.Client.GetKubeClient(), k8sComponents, ownerReference)
+	if err != nil {
+		return err
 	}
 
 	parameters.EnvSpecificInfo.SetDevfileObj(a.Devfile)